requirements.txt
flask
pandas
google-generativeai
python-dotenv
langchain
langchain-community
langchain_google_genai